requirements.txt
streamlit
google-generativeai
sentence-transformers
numpy
requests
torch